data for a new, simpler  version of the test
- first, the new test is meant to avoid measuring right/wrong
answers. rather, we'll simply measure preference: which axis
better represents the current audio? both cards will be “correct”,
so we'll be measuring a simple % of preference.

- second, although it's not yet implemented in the Evaluation.vue
component, I want to simplify the test organizing code and to
complicate a bit my data, in order to make it easier to change the
tests as results come in.
the idea is to stop shuffling card combinations. rather, i'll have
an array of cards and another array of combinations. the component
is supposed to shuffle these combinations (not all users will
necessarily go through all cards, so i want there to be a nice
distribution among them while also avoiding systematic errors
that a given order could induce), but it will stop knowing what
combinations work or not. no longer will it be necessary for it
to know that you can't mix different phrases, for instance.
this will mean some additional work because i'll have to hard code
these combinations, but this gives me flexibility to quickly add
or remove cards or combinations.
this will of course break compability with the previous cards/test
i had, so i made it a release